Refactor 40-0247 data-of-death custom-validation
