+++ src/applications/simple-forms/40-0247/pages/veteranPersonalInfo.js
@@ -1,7 +1,5 @@
 import React from 'react';
 
-import moment from 'moment';
-
 import {
   dateOfBirthUI,
   dateOfDeathUI,
@@ -10,6 +8,8 @@ import {
   fullNameNoSuffixSchema,
   fullNameNoSuffixUI,
 } from 'platform/forms-system/src/js/web-component-patterns';
+
+import { dateOfDeathValidation } from '../helpers';
 
 /** @type {PageSchema} */
 export default {
@@ -24,20 +24,7 @@ export default {
     veteranDateOfDeath: dateOfDeathUI(),
     // can't move this validation to Forms-Library web-component-patterns
     // since one pattern can't detect presence of the other
-    'ui:validations': [
-      // date of death should be after date of birth
-      (errors, fields) => {
-        const { veteranDateOfBirth, veteranDateOfDeath } = fields;
-        const dob = moment(veteranDateOfBirth);
-        const dod = moment(veteranDateOfDeath);
-
-        if (dod.isBefore(dob)) {
-          errors.veteranDateOfDeath.addError(
-            'Provide a date that is after the date of birth',
-          );
-        }
-      },
-    ],
+    'ui:validations': [dateOfDeathValidation],
   },
   schema: {
     type: 'object',
